+++ source/rafcon/core/states/state.py
@@ -139,9 +139,6 @@ class State(Observable, YAMLObject, JSONObject, Hashable):
     # ----------------------------------- generic methods -----------------------------------------
     # ---------------------------------------------------------------------------------------------
 
-    def __del__(self):
-        self._parent = None
-
     def __str__(self):
         return "{2} with name '{0}' and id '{1}'".format(self.name, self.state_id, type(self).__name__)
 
@@ -388,6 +385,7 @@ class State(Observable, YAMLObject, JSONObject, Hashable):
         # Check for name uniqueness
         valid, message = self._check_data_port_name(self._input_data_ports[data_port_id])
         if not valid:
+            self._input_data_ports[data_port_id].parent = None
             del self._input_data_ports[data_port_id]
             raise ValueError(message)
 
@@ -405,6 +403,7 @@ class State(Observable, YAMLObject, JSONObject, Hashable):
         if data_port_id in self._input_data_ports:
             if destroy:
                 self.remove_data_flows_with_data_port_id(data_port_id)
+            self._input_data_ports[data_port_id].parent = None
             return self._input_data_ports.pop(data_port_id)
         else:
             raise AttributeError("input data port with name %s does not exit", data_port_id)
@@ -450,6 +449,7 @@ class State(Observable, YAMLObject, JSONObject, Hashable):
         # Check for name uniqueness
         valid, message = self._check_data_port_name(self._output_data_ports[data_port_id])
         if not valid:
+            self._output_data_ports[data_port_id].parent = None
             del self._output_data_ports[data_port_id]
             raise ValueError(message)
 
@@ -466,6 +466,7 @@ class State(Observable, YAMLObject, JSONObject, Hashable):
         if data_port_id in self._output_data_ports:
             if destroy:
                 self.remove_data_flows_with_data_port_id(data_port_id)
+            self._output_data_ports[data_port_id].parent = None
             return self._output_data_ports.pop(data_port_id)
         else:
             raise AttributeError("output data port with name %s does not exit", data_port_id)
@@ -676,6 +677,7 @@ class State(Observable, YAMLObject, JSONObject, Hashable):
                     break  # found the one outgoing transition
 
         # delete outcome it self
+        self._outcomes[outcome_id].parent = None
         return self._outcomes.pop(outcome_id)
 
     @lock_state_machine
@@ -1043,6 +1045,11 @@ class State(Observable, YAMLObject, JSONObject, Hashable):
                 self._input_data_ports = old_input_data_ports
                 raise
 
+        # check that all old_input_data_ports are no more referencing self as there parent
+        for old_input_data_port in old_input_data_ports.itervalues():
+            if old_input_data_port not in self._input_data_ports.itervalues() and old_input_data_port.parent is self:
+                old_input_data_port.parent = None
+
     @property
     def output_data_ports(self):
         """Property for the _output_data_ports field
@@ -1096,6 +1103,11 @@ class State(Observable, YAMLObject, JSONObject, Hashable):
                 self._output_data_ports = old_output_data_ports
                 raise
 
+        # check that all old_output_data_ports are no more referencing self as there parent
+        for old_output_data_port in old_output_data_ports.itervalues():
+            if old_output_data_port not in self._output_data_ports.itervalues() and old_output_data_port.parent is self:
+                old_output_data_port.parent = None
+
     @property
     def outcomes(self):
         """Property for the _outcomes field
@@ -1145,6 +1157,11 @@ class State(Observable, YAMLObject, JSONObject, Hashable):
             self._outcomes[-1] = Outcome(outcome_id=-1, name="aborted", parent=self)
         if -2 not in outcomes:
             self._outcomes[-2] = Outcome(outcome_id=-2, name="preempted", parent=self)
+
+        # check that all old_outcomes are no more referencing self as there parent
+        for old_outcome in old_outcomes.itervalues():
+            if old_outcome not in self._outcomes.itervalues() and old_outcome.parent is self:
+                old_outcome.parent = None
 
     @property
     def input_data(self):